Added a file logger and a database logger

+++ Phoneshop.Business/PhoneService.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Phoneshop.Domain.Interfaces;
-using Phoneshop.Domain.Objects;
+using Phoneshop.Domain.Entities;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
-
 
 namespace Phoneshop.Business
 {
@@ -12,11 +12,13 @@ namespace Phoneshop.Business
     {
         private readonly IBrandService _brandService;
         private readonly IRepository<Phone> _phoneRepository;
+        private readonly ILogger _logger;
 
-        public PhoneService(IBrandService brandService, IRepository<Phone> phoneRepository)
+        public PhoneService(IBrandService brandService, IRepository<Phone> phoneRepository, ILogger logger)
         {
             _brandService = brandService;
             _phoneRepository = phoneRepository;
+            _logger = logger;
         }
 
         private IQueryable<Phone> Phones()
@@ -26,6 +28,14 @@ namespace Phoneshop.Business
 
         public Phone Get(int id)
         {
+            if (id <= 0)
+            {
+                Exception ex = new ArgumentOutOfRangeException(nameof(id));
+
+                _logger.Error(ex);
+                throw ex;
+            }
+
             var phone = Phones().SingleOrDefault(x => x.Id == id);
 
             return phone;
@@ -38,6 +48,8 @@ namespace Phoneshop.Business
 
         public IEnumerable<Phone> Search(string query)
         {
+            _logger.Info($"Search in list for: {query}");
+
             return Phones().Where(x => x.Brand.Name.Contains(query)
                                         || x.Type.Contains(query)
                                         || x.Description.Contains(query));
@@ -45,7 +57,17 @@ namespace Phoneshop.Business
 
         public void Delete(int id)
         {
+            if (id <= 0)
+            {
+                Exception ex = new ArgumentOutOfRangeException(nameof(id));
+
+                _logger.Error(ex);
+                throw ex;
+            }
+
             _phoneRepository.Delete(id);
+
+            _logger.Info($"Deleted a phone with id: {id}");
         }
 
         public void Create(Phone phone)
@@ -71,8 +93,12 @@ namespace Phoneshop.Business
             List<Brand> updatedBrandList = _brandService.GetAll().ToList();
             var brandItem = updatedBrandList.Find(x => x.Name.ToLower() == phone.Brand.Name.ToLower());
             phone.BrandID = brandItem.Id;
+            var name = phone.FullName;
+            phone.Brand = null;
 
             _phoneRepository.Create(phone);
+
+            _logger.Info($"Created a new phone: {name}");
         }
     }
 }
